forced type conversion to float from np.float for weight norms

+++ scripts/train-algorithmic-transformer.py
@@ -283,9 +283,9 @@ def run(operators,
                     embedding_l2 = sum(torch.pow(p, 2).sum() for p in model.embedding.parameters())
                     decoder_l2 = sum(torch.pow(p, 2).sum() for p in model.decoder.parameters())
                     last_layer_l2 = sum(torch.pow(p, 2).sum() for p in model.linear.parameters())
-                    ex.info['l2'].append(np.sqrt(embedding_l2.item() + decoder_l2.item() + last_layer_l2.item()))
-                    ex.info['decoder_l2'].append(np.sqrt(decoder_l2.item() + last_layer_l2.item()))
-                    ex.info['last_layer_l2'].append(np.sqrt(last_layer_l2.item()))
+                    ex.info['l2'].append(float(np.sqrt(embedding_l2.item() + decoder_l2.item() + last_layer_l2.item())))
+                    ex.info['decoder_l2'].append(float(np.sqrt(decoder_l2.item() + last_layer_l2.item())))
+                    ex.info['last_layer_l2'].append(float(np.sqrt(last_layer_l2.item())))
                 ex.info['log_steps'].append(steps)
                 if stop_early and ex.info['total']['val']['accuracy'][-1] > 0.98:
                     return
